chore: cross compile
